@@ -12,37 +12,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.2.4",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^4.0.5",
-    "@ngrx/store": "^4.0.3",
-    "@ngrx/store-devtools": "^4.0.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/http": "^7.2.16",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ngrx/effects": "^6.0.0",
+    "@ngrx/store": "^6.0.0",
+    "@ngrx/store-devtools": "^6.0.0",
     "angular2-sails": "^0.2.8",
     "angular2-svg-icons": "^1.1.1",
     "bootstrap": "^4.0.0-beta",
     "core-js": "^2.4.1",
     "ng-pwa-tools": "0.0.15",
     "ng2-dragula": "^1.5.0",
-    "ngrx-store-localstorage": "^0.2.2",
+    "ngrx-store-localstorage": "^6.0.0",
     "reselect": "^3.0.1",
-    "rxjs": "^5.4.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "6.6.3",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.4",
-    "@angular/compiler-cli": "^4.2.4",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular-devkit/core": "^0.5.12",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
     "@angular/language-service": "^4.2.4",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
+    "@types/socket.io-client": "^1.4.34",
     "codelyzer": "~3.2.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
@@ -52,9 +56,9 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ngrx-store-freeze": "^0.1.9",
     "prettier": "^1.7.0",
     "protractor": "~5.1.2",
+    "socket.io-client": "^3.0.1",
     "stylelint": "^8.1.1",
     "stylelint-config-sass-guidelines": "^3.1.0",
     "stylelint-scss": "^2.1.0",
@@ -62,6 +66,6 @@
     "tslint": "~5.7.0",
     "tslint-config-prettier": "^1.5.0",
     "tslint-config-standard": "^6.0.1",
-    "typescript": "~2.3.3"
+    "typescript": "3.5.3"
   }
 }
