@@ -1,12 +1,9 @@
-import "rxjs/add/operator/catch";
-import "rxjs/add/operator/debounceTime";
-import "rxjs/add/operator/map";
-import "rxjs/add/operator/switchMap";
+import { catchError, switchMap, map, debounceTime } from "rxjs/operators";
 
 import * as checklistActions from "./checklist.actions";
 import * as fromRoot from "../app.reducer";
 
-import { Actions, Effect, toPayload } from "@ngrx/effects";
+import { Actions, Effect } from "@ngrx/effects";
 import {
   ChecklistBlock,
   ChecklistItem,
@@ -16,10 +13,8 @@ import { Inject, Injectable, InjectionToken, Optional } from "@angular/core";
 
 import { Action } from "@ngrx/store";
 import { ChecklistService } from "./checklist.service";
-import { Observable } from "rxjs/Observable";
+import { Observable, EMPTY, of } from "rxjs";
 import { Store } from "@ngrx/store";
-import { empty } from "rxjs/observable/empty";
-import { of } from "rxjs/observable/of";
 
 export const USER_DEBOUNCE = new InjectionToken<number>("User Debounce");
 
@@ -28,44 +23,52 @@ export class ChecklistEffect {
   @Effect()
   loadChecklist$: Observable<Action> = this.actions$
     .ofType(checklistActions.LOAD_CHECKLIST)
-    .debounceTime(this.debounce)
-    .map(toPayload)
-    .switchMap(id => {
-      // if there are no user credentails
-      if (!id) {
-        return empty();
-      }
+    .pipe(
+      debounceTime(this.debounce),
+      switchMap((data: any) => {
+        // if there are no user credentails
+        const id = data.payload;
+        if (!id) {
+          return EMPTY;
+        }
 
-      return this._checklistService
-        .loadChecklist(id)
-        .map(fromBackend => {
-          const checklist: ChecklistModel = fromBackend.response.data.checklist;
-          return new checklistActions.SetChecklistAction(checklist);
-        })
-        .catch(() => of(new checklistActions.SetChecklistAction(null)));
-    });
+        return this._checklistService.loadChecklist(id).pipe(
+          map(fromBackend => {
+            const checklist: ChecklistModel =
+              fromBackend.response.data.checklist;
+            return new checklistActions.SetChecklistAction(checklist);
+          }),
+          catchError(() => of(new checklistActions.SetChecklistAction(null)))
+        );
+      })
+    );
 
   @Effect()
   postChecklist$: Observable<Action> = this.actions$
     .ofType(checklistActions.API_POST_CHECKLIST)
-    .debounceTime(this.debounce)
-    .map(toPayload)
-    .switchMap(checklist => {
-      // if there are no user credentails
-      if (!checklist) {
-        return empty();
-      }
+    .pipe(
+      debounceTime(this.debounce),
+      switchMap((data: any) => {
+        const checklist = data.payload;
+        // if there are no user credentails
+        if (!checklist) {
+          return EMPTY;
+        }
 
-      return this._checklistService
-        .createChecklist(checklist)
-        .map(fromBackend => {
-          const checklist: ChecklistModel = fromBackend.response.data.checklist;
+        return this._checklistService.createChecklist(checklist).pipe(
+          map((fromBackend: any) => {
+            const checklist: ChecklistModel =
+              fromBackend.response.data.checklist;
 
-          this._store.dispatch(new checklistActions.SetChecklistChangeAction());
-          return new checklistActions.SetChecklistAction(checklist);
-        })
-        .catch(() => of(new checklistActions.SetChecklistAction(null)));
-    });
+            this._store.dispatch(
+              new checklistActions.SetChecklistChangeAction()
+            );
+            return new checklistActions.SetChecklistAction(checklist);
+          }),
+          catchError(() => of(new checklistActions.SetChecklistAction(null)))
+        );
+      })
+    );
 
   constructor(
     private actions$: Actions,
