@@ -1,31 +1,31 @@
-import * as checklistActions from '../checklist-service/checklist.actions';
-import * as fromRoot from '../app.reducer';
-import * as userActions from '../shared/user/user.actions';
+import { first, filter, catchError, map } from "rxjs/operators";
+import * as checklistActions from "../checklist-service/checklist.actions";
+import * as fromRoot from "../app.reducer";
+import * as userActions from "../shared/user/user.actions";
 
-import { ElementRef, Injectable, Renderer } from '@angular/core';
+import { ElementRef, Injectable, Renderer } from "@angular/core";
 
-import { ChecklistModel } from './checklist.interface';
-import { ErrorHandlingService } from '../shared/error-handling/error-handling.service';
-import { Http } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-import { RestService } from '../shared/rest/rest.service';
-import { SailsService } from 'angular2-sails';
-import { StatusBarService } from '../status-bar/status-bar.service';
-import { Store } from '@ngrx/store';
-import { Subject } from 'rxjs/Subject';
-import { defaultChecklist } from './default-checklist';
-import { environment } from '../../environments/environment';
+import { ChecklistModel } from "./checklist.interface";
+import { ErrorHandlingService } from "../shared/error-handling/error-handling.service";
+import { HttpClient, HttpHeaders } from "@angular/common/http";
+import { Observable, Subject } from "rxjs";
+import { RestService } from "../shared/rest/rest.service";
+import { SailsService } from "angular2-sails";
+import { StatusBarService } from "../status-bar/status-bar.service";
+import { Store } from "@ngrx/store";
+import { defaultChecklist } from "./default-checklist";
+import { environment } from "../../environments/environment";
 
 @Injectable()
 export class ChecklistService {
   public appEl: any;
   public appRenderer: any;
-  private headers;
+  private headers: HttpHeaders;
   public scrollData = new Subject();
-  public inputToFocus = '';
+  public inputToFocus = "";
 
   constructor(
-    private _http: Http,
+    private _http: HttpClient,
     private _store: Store<fromRoot.State>,
     private _statusBarService: StatusBarService,
     private _restService: RestService,
@@ -40,7 +40,7 @@ export class ChecklistService {
   }
 
   isMobile() {
-    if (typeof navigator !== 'undefined') {
+    if (typeof navigator !== "undefined") {
       return (
         navigator.userAgent.match(
           /Android|webOS|iPhone|iPad|iPod|BlackBerry|Windows Phone/i
@@ -93,11 +93,11 @@ export class ChecklistService {
 
     checklists.splice(15, 1);
 
-    localStorage.setItem('latest-checklists', JSON.stringify(checklists));
+    localStorage.setItem("latest-checklists", JSON.stringify(checklists));
   }
 
   getLatestFromLocalStore() {
-    const checklists = localStorage.getItem('latest-checklists');
+    const checklists = localStorage.getItem("latest-checklists");
     if (checklists) {
       return JSON.parse(checklists);
     } else {
@@ -116,8 +116,12 @@ export class ChecklistService {
       this._store.dispatch(new checklistActions.SetChecklistAction(checklist));
       this._store
         .select(fromRoot.getUser)
-        .filter(user => !!user)
-        .first()
+        .pipe(
+          filter(user => {
+            return !!user;
+          }),
+          first()
+        )
         .subscribe(user => {
           this.addChecklistToUser(user.id, checklist);
         });
@@ -126,38 +130,43 @@ export class ChecklistService {
 
   loadChecklist(id) {
     return this._http
-      .get(`${environment.sailsApi}/checklist/${id}`, this.headers)
-      .map(res => {
-        try {
-          // ensure that there cannot be any brackets '<' '>' in the json
-          const cleanedResult = res.text().replace(/<\/?[^>]+(>|$)/g, '');
-          return JSON.parse(cleanedResult) || {};
-        } catch (err) {
-          this._statusBarService.setStatus(err, 'error');
-        }
-      })
-      .catch(e => {
-        return this._errorHandlingService.errorHandler(e);
-      });
+      .get(`${environment.sailsApi}/checklist/${id}`, { headers: this.headers })
+      .pipe(
+        map(res => {
+          try {
+            // ensure that there cannot be any brackets '<' '>' in the json
+            const cleanedResult = JSON.stringify(res).replace(
+              /<\/?[^>]+(>|$)/g,
+              ""
+            );
+            return JSON.parse(cleanedResult) || {};
+          } catch (err) {
+            this._statusBarService.setStatus(err, "error");
+          }
+        }),
+        catchError(e => {
+          return this._errorHandlingService.errorHandler(e);
+        })
+      );
   }
 
   createChecklist(checklist) {
     return this._http
-      .post(
-        `${environment.sailsApi}/checklist`,
-        JSON.stringify(checklist),
-        this.headers
-      )
-      .map(res => {
-        try {
-          return res.json() || {};
-        } catch (err) {
-          this._statusBarService.setStatus(err, 'error');
-        }
+      .post(`${environment.sailsApi}/checklist`, JSON.stringify(checklist), {
+        headers: this.headers
       })
-      .catch(e => {
-        return this._errorHandlingService.errorHandler(e);
-      });
+      .pipe(
+        map(res => {
+          try {
+            return res;
+          } catch (err) {
+            this._statusBarService.setStatus(err, "error");
+          }
+        }),
+        catchError(e => {
+          return this._errorHandlingService.errorHandler(e);
+        })
+      );
   }
 
   saveChecklist(
@@ -191,36 +200,21 @@ export class ChecklistService {
       this._store.dispatch(changeAction);
     }
 
-    /*     return this._http
-      .put(
-        `${environment.sailsApi}/checklist/${id}`,
-        JSON.stringify(checklist),
-        this.headers
-      )
-      .map(res => {
-        try {
-          return res.json() || {};
-        } catch (err) {
-          this._statusBarService.setStatus(err, 'error');
-        }
-      })
-      .catch(e => {
-        return this._errorHandlingService.errorHandler(e);
-      }); */
-
     // save via socket
 
     return this._sailsService
       .put(`${environment.sailsApi}/checklist/${id}`, checklist)
-      .map(res => {
-        try {
-          return res.data.response.data.checklist[0] || {};
-        } catch (err) {
-          this._statusBarService.setStatus(err, 'error');
-        }
-      })
-      .catch(e => {
-        return this._errorHandlingService.errorHandler(e);
-      });
+      .pipe(
+        map(res => {
+          try {
+            return res.data.response.data.checklist[0] || {};
+          } catch (err) {
+            this._statusBarService.setStatus(err, "error");
+          }
+        }),
+        catchError(e => {
+          return this._errorHandlingService.errorHandler(e);
+        })
+      );
   }
 }
